+++ src/layout/crossingReducer.ts
@@ -1,0 +1,5 @@
+import { PedigreeGraph } from "../model/pedigreeGraph";
+
+export function reduceCrossings(graph: PedigreeGraph) {
+  return graph;
+}