+++ src/model/pedigreeGraph.ts
@@ -1,0 +1,8 @@
+import { Person } from "./person";
+import { UnionNode } from "./union";
+
+export interface PedigreeGraph {
+  persons: Map<string, Person>;
+  unions: Map<string, UnionNode>;
+  childrenMap: Map<string, string[]>;
+}